requirements.txt
Flask == 1.1.2
gunicorn == 20.1.0
itsdangerous==2.0.1
Jinja2==2.11.3
MarkupSafe==2.0.1
Werkzeug==2.0.3
numpy>=1.9.2
scipy>=0.15.1
scikit-learn>=0.18
matplotlib>=1.4.3
pandas>=0.19
requests==2.28.1
mysql-connector-python==8.1.0


